pyproject.toml
[project]
name = "calm-lms"
version = "0.1.0"
description = ""
authors = [
    {name = "Aditya",email = "adityaburgula123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework (>=3.15.2,<4.0.0)",
    "firebase-admin (>=6.6.0,<7.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-cors-headers (>=4.6.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
